requirements.txt
bottle==0.12.19
bottle-swagger-2==2.0.11
meinheld==1.0.2
PyYAML==5.4.1
black==19.10b0
isort==4.3.21
flake8==3.7.9
pre-commit==1.21.0
transformers==4.6.1
scikit-learn==0.24.2
networkx==2.5.1
sentence-transformers==1.2.0
protobuf==3.17.1
